+++ CHANGELOG.md
@@ -3,3 +3,7 @@
  - Added a pairing guide (replace the access denied page)
  - Added a configuration guide on top of the settings page
  - Added a "prepared" state triggered when the shipping label of an order is created
+
+ # 1.1.1
+  - Selected parcel point is now displayed in admin order details
+  - Selected parcel point is now displayed in client order details